web/flask-dashboard-adminkit/pyproject.toml
[tool.poetry]
name = "flask-empty"
version = "0.1.0"
description = ""
authors = ["Sergey Fomchukov <durovam06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.1"
flask-login = "^0.6.3"
flask-bcrypt = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
flask-wtf = "^1.2.2"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
